requirements.txt
crewai>=0.28.0
streamlit>=1.25.0
pandas>=2.0.0
pyyaml>=6.0
openai>=1.0.0
anthropic>=0.5.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0
